added `Replies` and `MessageReaction` structs. added some fields in `MessageContainer` and `QuotedMessage`. fixed `Message.from_protobuf`

+++ balethon/objects/location.py
@@ -5,10 +5,10 @@ class Location(Object):
 
     def __init__(
             self,
-            longitude: int = None,
-            latitude: int = None,
+            longitude: float = None,
+            latitude: float = None,
             **kwargs
     ):
         super().__init__(**kwargs)
-        self.longitude: int = longitude
-        self.latitude: int = latitude
+        self.longitude: float = longitude
+        self.latitude: float = latitude
